"Update Gradle and Kotlin versions, reorganize code for the new year"
This commit updates the Gradle version from 7.6 to 8.4 and the Kotlin version from 1.7.20 to 1.9.21. These updates were made to take advantage of new functionalities and improvements in the newer versions of these tools.

In addition, the project's code structure has been modified. The code files for the days of the AdventOfCode challenge have been renamed and moved from the main directory to subdirectories dedicated to each year (for instance 2022 and 2023). This change in structure is designed to make the code organization cleaner and easier to navigate, especially as more daily challenges are added in the future.

Several new utility files for the years 2022 and 2023 were also added to support the updated code organization. The changes to how resources are loaded are also reflected in the code of individual daily challenges.

A new file called ".idea/gradle.xml" has been added to the .idea directory, which appears to be related to migration settings for Gradle.

Moreover, the "settings.gradle.kts" file appears to be cleaned up (removing empty lines) and several project configuration files in the .idea directory have been updated. These changes seem to be automatically generated by the IDE and are likely related to the updates to Gradle and Kotlin versions.

Lastly, it removes an unused temporary Gradle run configuration from the ".idea/workspace.xml" file. This cleanup makes the project settings more maintainable and can potentially prevent future issues caused by outdated or erroneous configurations.